Add Venti3D dataset configuration and processing scripts
- Created dataset configuration file for Venti3D dataset with data paths, point cloud range, and data augmentations.
- Added model configuration files for Lion Mamba with different dimensions and class names.
- Implemented data processing script to create dataset info and ground truth database for Venti3D, including class mapping and point cloud handling.

+++ pcdet/models/backbones_3d/lion_backbone_one_stride.py
@@ -368,7 +368,7 @@ class LIONLayer(nn.Module):
 
             x_features = block(x_features)
 
-            x.features[indices] = x_features.view(-1, x_features.shape[-1])[mappings["win2flat"]]
+            x.features[indices] = x_features.view(-1, x_features.shape[-1])[mappings["win2flat"]].to(x.features.dtype)
 
         return x
